% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{CreateKnmiTFInput}
\alias{CreateKnmiTFInput}
\title{Create valid input from user specified data}
\usage{
CreateKnmiTFInput(data, stationID, lat, lon, comment = NULL)
}
\arguments{
\item{data}{a data.frame of two columns
\describe{
\item{date}{integer vector date format yyyymmdd }
\item{values}{numeric values of observation}
}}

\item{stationID}{numeric (either official station number or self-chosen /
recommendations?)}

\item{lat}{numeric latitude}

\item{lon}{numeric longitude}

\item{comment}{user specified comment (should start with '#') default is the
current timestamp}
}
\description{
can be used if one does not have the KNMI standard format
}
\note{
Currently for single stations only
}
