% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationTransformation.R
\name{daynumber}
\alias{daynumber}
\title{Derive daynumber within year (1-366)}
\usage{
daynumber(datestring)
}
\arguments{
\item{datestring}{integer in format YYYYMMDD}
}
\description{
Derive daynumber within year (1-366)
}
\keyword{internal}
