% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droogte_transformatie_KNMI14.R
\name{droogte_berekening_KNMI14}
\alias{droogte_berekening_KNMI14}
\title{Calculation of Makkink evaporation}
\usage{
droogte_berekening_KNMI14(ifile_tg, ifile_rsds, ofile = "uitvoer.txt", sc,
  p = NA, regio.file = NA)
}
\arguments{
\item{ifile_tg}{Name of the input file for temperature}

\item{ifile_rsds}{Name of the input file for radiation}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to ifile}

\item{sc}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{p}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{regio.file}{this (optional) argument provides the name of an ASCII file that relates the stations to
a particular region. First column is station id and second column region
KNMI14 distinguishes following regions:
<NLD> Nederland            [DEFAULT]
<NWN> Noordwest Nederland
<ZWN> Zuidwest Nederland
<NON> Noordoost Nederland
<MON> Middenoost Nederland
<ZON> Zuidoost Nederland}
}
\description{
Function reads transormed mean temperature and transformed global radiation
and calculates the Makkink evaporation for 'future time series' that match a certain climate
}

