% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{MatchRegionsOnStationId}
\alias{MatchRegionsOnStationId}
\title{Match regions on station id}
\usage{
MatchRegionsOnStationId(stationId)
}
\arguments{
\item{stationId}{Station id}
}
\value{
vector with regions
}
\description{
Match regions on station id
}
