% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neerslag_transformatie_KNMI14.R
\name{neerslag_transformatie_KNMI14}
\alias{neerslag_transformatie_KNMI14}
\title{Transformation of precipitation}
\usage{
neerslag_transformatie_KNMI14(ifile, ofile = "uitvoer.txt", delta.file = NA,
  sc, p = 2030, scaling = "centr", version = "v1.1")
}
\arguments{
\item{ifile}{Name of the input file (ASCII) that contains reference data (all numerics) in which
the columns provide time series for specific stations. The first column
should provide either 00000000 or a datestring YYYYMMDD:
Rows starting with 00000000 are considered station info (station number, lat, lon
etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the time series.
Rows starting with "#" are completely ignored and returned unchanged}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to ifile}

\item{delta.file}{[optional] Name of file that contains deltas (changes factors for the transformation)
               File should contain following compulsory columns identified with compulsory headers
               "wdf"       relative change in wet-day frequency
                           (wet days are defined as days with 0.1 mm or more precipitation) \cr
               "ave"       relative change in mean precipitation \cr
               "p99.lower" lower   estimate of the relative change in the 99th percentile of wet-day amounts \cr
               "p99.centr" central estimate of the relative change in the 99th percentile of wet-day amounts \cr
               "p99.upper" upper   estimate of the relative change in the 99th percentile of wet-day amounts \cr

                (If delta.file is not provided, predefined deltas are derived dependening on <sc>, <p> and <scaling>)}

\item{sc}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{p}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{scaling}{scaling extreme precipitation ["lower", "centr" (=DEFAULT), "upper"]}

\item{version}{"v1.1" [DEFAULT] official version that belongs to KNMI'14}
}
\description{
Function reads 'reference data' with daily precipitation sums
[mm] and 'change factors' from input files and applies them to function
'rr_trans_KNMI14' to obtain 'future time series' that match a certain climate
}

