% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationTransformation.R
\name{ObtainAngotRadiation}
\alias{ObtainAngotRadiation}
\title{Obtain Top of Atmosphere radiation}
\usage{
ObtainAngotRadiation(date, lat)
}
\arguments{
\item{date}{datestring format yyyymmdd}

\item{lat}{latitude}
}
\description{
Obtain Top of Atmosphere radiation
}
\keyword{internal}
