% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{ObtainMonth}
\alias{ObtainMonth}
\title{Obtains corresponding month}
\usage{
ObtainMonth(day)
}
\arguments{
\item{day}{integer in format yyyymmdd}
}
\description{
Obtains corresponding month
}
\keyword{internal}
