% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{ReadInput}
\alias{ReadInput}
\title{Reads knmi standard file}
\usage{
ReadInput(var, ifile)
}
\arguments{
\item{var}{type \code{[rr, evmk, rsds, tg, tn, tx]}}

\item{ifile}{filename}
}
\value{
list with comments, header and observations
}
\description{
Reads knmi standard file
}
