% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationTransformation.R
\name{rsds_trans_KNMI14}
\alias{rsds_trans_KNMI14}
\title{rsds_trans_KNMI14}
\usage{
rsds_trans_KNMI14(obs, deltas, lat)
}
\arguments{
\item{obs}{data.frame or matrix: \cr
first column provides datestring YYYYMMDD \cr
other columns provide global radiation (kJ/m2) time series
(each column represents specific station)}

\item{deltas}{data.frame or matrix that contains deltas (=change factors for the transformation)
should contain following columns indicated by following headers \cr
HEADER \cr
"ave"       relative change (\%) in mean shortwave surface radiation}

\item{lat}{vector with latitudes belonging to stations.
location (row.number (i) in vector refers to column in obs (i+1)}
}
\description{
Function 'transforms' a specific reference-dataset with time
series of daily global radiation sums (kJ/m2) to a dataset representative
for a future climate scenario.
}
\keyword{internal}
