% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitReferenceDataSets.R
\name{SplitReferenceFile}
\alias{SplitReferenceFile}
\title{Split reference data file into single files}
\usage{
SplitReferenceFile(filename)
}
\description{
Split reference data file into single files
}

