% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straling_transformatie_KNMI14.R
\name{straling_transformatie_KNMI14}
\alias{straling_transformatie_KNMI14}
\title{Transformation of radiation}
\usage{
straling_transformatie_KNMI14(ifile, ofile = "uitvoer.txt", delta.file = NA,
  sc, p = NA)
}
\arguments{
\item{ifile}{Name of the input file (ASCII) that contains reference data
(all numerics) in which the columns provide time series for
specific stations.
The first column should provide either 00000000 or a
datestring YYYYMMDD:
Rows starting with 00000000 are considered station info
(station number, lat, lon etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the
time series.
Rows starting with "#" are completely ignored and returned
unchanged.}

\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the
transformed data to.
Format is similar to ifile}

\item{delta.file}{[optional] Name of file that contains deltas (changes factors for the transformation)
File should contain following compulsory column identified with compulsory headers
- HEADER -
"ave"       relative change [\%] in mean shortwave surface radiation}

\item{sc}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{p}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}
}
\description{
Function reads 'reference data' with time series of daily global
 radiation sums [kJ/m2] and 'change factors' from input files and applies
 them to function 'rsds_trans_KNMI14' to obtain 'future time series' that
 match a certain climate
}
\note{
The 5th row of the ifile indicated by 00000000 must be given as it is
interpreted to contain LATITUDES of station.
}

