% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droogte_transformatie_KNMI14.R,
%   R/knmitransformer.R
\name{TransformEvap}
\alias{TransformEvap}
\alias{TransformEvap}
\title{Calculation of Makkink evaporation}
\usage{
TransformEvap(input_tg, input_rsds, scenario = "GL", horizon = 2030,
  regio.file = NA, ofile = NA, rounding = TRUE)

TransformEvap(input_tg, input_rsds, scenario = "GL", horizon = 2030,
  regio.file = NA, ofile = NA, rounding = TRUE)
}
\arguments{
\item{input_tg}{Name of the input file for temperature}

\item{input_rsds}{Name of the input file for radiation}

\item{scenario}{scenario ("GL"=DEFAULT, "GH", "WL", "WH")}

\item{horizon}{time horizon (2030=DEFAULT, 2050, 2085)}

\item{regio.file}{this (optional) argument provides the name of an ASCII file
that relates the stations to a particular region.
First column is station id and second column region
KNMI14 distinguishes following regions:
<NLD> Nederland            (DEFAULT)
<NWN> Noordwest Nederland
<ZWN> Zuidwest Nederland
<NON> Noordoost Nederland
<MON> Middenoost Nederland
<ZON> Zuidoost Nederland}

\item{ofile}{Name of the output file to write the transformed data to.
Format is similar to input file mentioned in input. By default no output
file is written.}

\item{rounding}{Logical (default = TRUE) if results should be rounded}

\item{ifile_tg}{Name of the input file for temperature}

\item{ifile_rsds}{Name of the input file for radiation}
}
\description{
Function reads transormed mean temperature and transformed global radiation
and calculates the Makkink evaporation for 'future time series' that match a certain climate

Function reads transformed mean temperature and transformed
global radiation and calculates the Makkink evaporation for 'future time
series' that match a certain climate
}
