% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neerslag_transformatie_KNMI14.R
\name{TransformPrecip}
\alias{TransformPrecip}
\title{Transformation of precipitation}
\usage{
TransformPrecip(ifile, ofile = NA, scenario, horizon = 2030,
  subscenario = "centr", rounding = TRUE)
}
\arguments{
\item{ifile}{Name of the input file (ASCII) that contains reference data
(all numerics) in which the columns provide time series for
specific stations.
The first column should provide either 00000000 or a
datestring YYYYMMDD:
Rows starting with 00000000 are considered station info
(station number, lat, lon etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the
time series.
Rows starting with "#" are completely ignored and returned
unchanged.}

\item{ofile}{(DEFAULT=NA) Name of the output file to write the
transformed data to.
Format is similar to ifile}

\item{scenario}{scenario ("GL", "GH", "WL", "WH")}

\item{horizon}{time horizon (2030=DEFAULT, 2050, 2085)}

\item{subscenario}{subscenario for extreme precipitation
("lower", "centr" (=DEFAULT), "upper")}

\item{rounding}{Logical (default = TRUE) if results should be rounded}
}
\description{
Function reads 'reference data' with daily precipitation sums
\code{[mm]} and 'change factors' from input files and applies them to function
'rr_trans_KNMI14' to obtain 'future time series' that match a certain climate
}
