% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmitransformer.R
\name{TransformTemp}
\alias{TransformTemp}
\title{Transformation of temperature}
\usage{
TransformTemp(ifile, ofile = NA, scenario, horizon = 2030, var,
  regio.file = NA, rounding = TRUE)
}
\arguments{
\item{ifile}{Name of the input file (ASCII) that contains reference data
(all numerics) in which the columns provide time series for
specific stations.
The first column should provide either 00000000 or a
datestring YYYYMMDD:
Rows starting with 00000000 are considered station info
(station number, lat, lon etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the
time series.
Rows starting with "#" are completely ignored and returned
unchanged.}

\item{ofile}{(DEFAULT=NA) Name of the output file to write the
transformed data to.
Format is similar to ifile}

\item{scenario}{scenario ("GL", "GH", "WL", "WH")}

\item{horizon}{time horizon (2030=DEFAULT, 2050, 2085)}

\item{var}{kind of daily temperature variable
("tg" = daily mean, "tn" = daily minimum, "tx" = daily maximum)}

\item{regio.file}{this (optional) argument provides the name of an ASCII file
that relates the stations to a particular region.
First column is station id and second column region
KNMI14 distinguishes following regions:
<NLD> Nederland            (DEFAULT)
<NWN> Noordwest Nederland
<ZWN> Zuidwest Nederland
<NON> Noordoost Nederland
<MON> Middenoost Nederland
<ZON> Zuidoost Nederland}

\item{rounding}{Logical (default = TRUE) if results should be rounded}
}
\description{
Function reads 'reference data' with time series of daily of
mean, minimum or maximum temperature \code{[degree Celsius]} and 'change factors'
from input files and applies them to function 'tm_trans_KNMI14' to obtain
'future time series' that match a certain climate
}
