% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droogte_transformatie_KNMI14.R
\name{droogte_berekening_KNMI14}
\alias{droogte_berekening_KNMI14}
\title{Calculation of Makkink evaporation}
\usage{
droogte_berekening_KNMI14(ifile_tg, ifile_rsds, ofile = "uitvoer.txt",
  delta.file.rsds, delta.file.tg = NA, sc, p = NA, regio.file = NA)
}
\arguments{
\item{ofile}{(DEFAULT="uitvoer.txt") Name of the output file to write the transformed data to.
Format is similar to ifile}

\item{delta.file.rsds}{[optional] Name of file that contains deltas (changes factors for the transformation)
File should contain following compulsory column identified with compulsory headers
- HEADER -
"ave"       relative change [\%] in mean shortwave surface radiation}

\item{delta.file.tg}{[optional] Name of file that contains deltas (changes factors for the transformation)
               File should contain following compulsory columns identified with compulsory headers
              - HEADER -
              "maand"     month for which deltas are valid (1,2,...,12) \cr
              "P01"       1st  percentile daily temperature \cr
              "P05"       5th  percentile daily temperature \cr
              "P50"       50th percentile daily temperature \cr
              "P95"       95th percentile daily temperature \cr

              following column is optional in case deltas vary with region
              (is needed in case <regio.tabel> is provided)
              "regio"     region for which deltas are valid
                          KNMI14 distinguishes ("NWN", "ZWN", "NON", "MON", "ZON", "NLD")

                (If delta.file is not provided, predefined deltas are derived dependening on <sc>, <p> and
                 daily temperature characteristic of interest [mean, min or max])}

\item{sc}{scenario                      ["GL", "GH", "WL", "WH"]}

\item{p}{time horizon                  [2030 (=DEFAULT), 2050, 2085]}

\item{regio.file}{this (optional) argument provides the name of an ASCII file that relates the stations to
a particular region. First column is station id and second column region
KNMI14 distinguishes following regions:
<NLD> Nederland            [DEFAULT]
<NWN> Noordwest Nederland
<ZWN> Zuidwest Nederland
<NON> Noordoost Nederland
<MON> Middenoost Nederland
<ZON> Zuidoost Nederland}

\item{ifile}{Name of the input file (ASCII) that contains reference data (all numerics) in which
the columns provide transformed time series for specific stations. The first column
should provide either 00000000 or a datestring YYYYMMDD:
Rows starting with 00000000 are considered station info (station number, lat, lon
etc.) and are ignored.
Rows starting with a datestring refer to a specific day in the time series.
Rows starting with "#" are completely ignored and returned unchanged}

\item{var}{daily Makkink evaporation [evmk= evaporation makkink]}
}
\description{
Function reads transormed mean temperature and transformed global radiation
and calculates the Makkink evaporation for 'future time series' that match a certain climate
}

