% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiationTransformation.R
\name{ObtainAngotRadiation}
\alias{ObtainAngotRadiation}
\title{Obtain Top of Atmossphere radiation}
\usage{
ObtainAngotRadiation(datestring, lat)
}
\arguments{
\item{datestring}{integer in format YYYYMMDD}

\item{lat}{latitude}
}
\description{
Obtain Top of Atmossphere radiation
}
\keyword{internal}
