% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_trans_KNMI14.R
\name{tm_trans_KNMI14}
\alias{tm_trans_KNMI14}
\title{tm_trans_KNMI14}
\usage{
tm_trans_KNMI14(obs, deltas, regio.tabel = NA)
}
\arguments{
\item{obs}{data.frame or matrix: \cr
first column provides datestring YYYYMMDD \cr
other columns provide precipitation [mm] time series (each column represents
specific station)}

\item{deltas}{data.frame or matrix that contains deltas (=change factors for the transformation)
               should contain following columns indicated by following headers
               HEADER \cr
              "maand"     month for which deltas are valid (1,2,...,12) \cr
              "P01"       1st  percentile daily temperature  \cr
              "P05"       5th  percentile daily temperature \cr
              "P50"       50th percentile daily temperature \cr
              "P95"       95th percentile daily temperature \cr
              "P99"       99th percentile daily temperature \cr

              following column is optional in case deltas vary with region
              (is needed in case <regio.tabel> is provided)
              "regio"     region for which deltas are valid
                          KNMI14 distinguishes ("NWN", "ZWN", "NON", "MON", "ZON", "NLD")}

\item{regio.tabel}{this (optional) argument provides a vector that relates the stations to
              a particular region. Vector provides references to region. Location (rownumber [i]
              in vector refers to column in obs [i+1]

              KNMI14 distinguishes following regions:
              <NLD> Nederland            [DEFAULT]
              <NWN> Noordwest Nederland
              <ZWN> Zuidwest Nederland
              <NON> Noordoost Nederland
              <MON> Middenoost Nederland
              <ZON> Zuidoost Nederland}
}
\description{
Function 'transforms' a specific reference-dataset with time
 series of daily of mean, minimum or maximum temperature [degrees Celsius] to
 a dataset representative for a future climate scenario.
}

