% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmitransformer.R
\name{TransformEvap}
\alias{TransformEvap}
\title{Calculation of Makkink evaporation}
\usage{
TransformEvap(inputTemp, inputRad, scenario = "GL", horizon = 2030,
  regions = "NLD", ofile = NA, rounding = TRUE)
}
\arguments{
\item{inputTemp}{Name of the input file for temperature}

\item{inputRad}{Name of the input file for radiation}

\item{scenario}{scenario ("GL"=DEFAULT, "GH", "WL", "WH")}

\item{horizon}{time horizon (2030=DEFAULT, 2050, 2085)}

\item{regions}{vector of regions
KNMI14 distinguishes following regions:\cr
<NLD> Nederland (DEFAULT) \cr
<NWN> Noordwest Nederland \cr
<ZWN> Zuidwest Nederland \cr
<NON> Noordoost Nederland \cr
<MON> Middenoost Nederland \cr
<ZON> Zuidoost Nederland}

\item{ofile}{Name of the output file to write the transformed data to.
Format is similar to input file mentioned in input. By default no output
file is written.}

\item{rounding}{Logical (default = TRUE) if results should be rounded}
}
\description{
Function reads transformed mean temperature and transformed
global radiation and calculates the Makkink evaporation for 'future time
series' that match a certain climate
}
